feat(cli-vector): merge stac functionality
